added test to check not null string
