+++ app/src/main/AndroidManifest.xml
@@ -34,6 +34,8 @@
             android:name="com.google.android.gms.ads.AdActivity"
             android:configChanges="keyboard|keyboardHidden|orientation|screenLayout|uiMode|screenSize|smallestScreenSize"
             android:theme="@android:style/Theme.Translucent" />
+        <uses-library android:name="android.test.runner"/>
+
     </application>
 
 </manifest>
