+++ app/src/main/res/values/strings.xml
@@ -3,6 +3,5 @@
     <string name="instructions">Press the button for a delicious joke!</string>
     <string name="button_text">Tell Joke</string>
     <string name="action_settings">Settings</string>
-    <string name="base_url">https://builditbigger-143808.appspot.com/_ah/api/myApi/v1/</string>
     <string name="banner_ad_unit_id">ca-app-pub-3940256099942544/6300978111</string>
 </resources>
